@@ -1,8 +1,11 @@
-// swift-tools-version:4.2
+// swift-tools-version:5.0
 import PackageDescription
 
 let package = Package(
     name: "Nimble",
+    platforms: [
+      .macOS(.v10_10), .iOS(.v8), .tvOS(.v9)
+    ],
     products: [
         .library(name: "Nimble", targets: ["Nimble"]),
     ],
@@ -17,5 +20,5 @@ let package = Package(
             exclude: ["objc"]
         ),
     ],
-    swiftLanguageVersions: [.v4_2]
+    swiftLanguageVersions: [.v5]
 )
