@@ -1,12 +1,7 @@
 import Foundation
 
 internal func identityAsString(_ value: Any?) -> String {
-    let anyObject: AnyObject?
-#if os(Linux) && !swift(>=4.1.50)
-    anyObject = value as? AnyObject
-#else
-    anyObject = value as AnyObject?
-#endif
+    let anyObject = value as AnyObject?
     if let value = anyObject {
         return NSString(format: "<%p>", unsafeBitCast(value, to: Int.self)).description
     } else {
@@ -121,13 +116,7 @@ extension String: TestOutputStringConvertible {
 
 extension Data: TestOutputStringConvertible {
     public var testDescription: String {
-        #if os(Linux)
-            // swiftlint:disable:next todo
-            // FIXME: Swift on Linux triggers a segfault when calling NSData's hash() (last checked on 03-11-16)
-            return "Data<length=\(count)>"
-        #else
-            return "Data<hash=\((self as NSData).hash),length=\(count)>"
-        #endif
+        return "Data<hash=\((self as NSData).hash),length=\(count)>"
     }
 }
 
