@@ -33,9 +33,7 @@ final class EqualTest: XCTestCase, XCTestCaseProvider {
         expect(array1).to(equal([1, 2, 3]))
         expect(array1).toNot(equal([1, 2] as [Int]))
 
-#if os(macOS) || os(iOS) || os(tvOS) || os(watchOS)
         expect(NSArray(array: [1, 2, 3])).to(equal(NSArray(array: [1, 2, 3])))
-#endif
 
         failsWithErrorMessage("expected to equal <[1, 2]>, got <[1, 2, 3]>") {
             expect([1, 2, 3]).to(equal([1, 2]))
@@ -119,10 +117,8 @@ final class EqualTest: XCTestCase, XCTestCaseProvider {
         expect(actual).to(equal(expected))
         expect(actual).toNot(equal(unexpected))
 
-#if os(macOS) || os(iOS) || os(tvOS) || os(watchOS)
         expect(NSDictionary(object: "bar", forKey: "foo" as NSString)).to(equal(["foo": "bar"]))
         expect(NSDictionary(object: "bar", forKey: "foo" as NSString) as? [String: String]).to(equal(expected))
-#endif
     }
 
     func testDataEquality() {
@@ -133,14 +129,8 @@ final class EqualTest: XCTestCase, XCTestCaseProvider {
         expect(actual).to(equal(expected))
         expect(actual).toNot(equal(unexpected))
 
-        #if os(Linux)
-            // swiftlint:disable:next todo
-            // FIXME: Swift on Linux triggers a segfault when calling NSData's hash() (last checked on 03-11)
-            let expectedErrorMessage = "expected to equal <Data<length=9>>, got <Data<length=6>>"
-        #else
-            let expectedErrorMessage = "expected to equal <Data<hash=92856895,length=9>>,"
+        let expectedErrorMessage = "expected to equal <Data<hash=92856895,length=9>>,"
                 + " got <Data<hash=114710658,length=6>>"
-        #endif
 
         failsWithErrorMessage(expectedErrorMessage) {
             expect(actual).to(equal(unexpected))
