@@ -51,10 +51,12 @@ public class NMBObjCBeCloseToMatcher: NSObject, NMBMatcher {
             return actualExpression() as? NMBDoubleConvertible
         })
         let expr = Expression(expression: actualBlock, location: location)
-        let matcher = beCloseTo(self._expected, within: self._delta)
+        let predicate = beCloseTo(self._expected, within: self._delta)
 
         do {
-            return try matcher.matches(expr, failureMessage: failureMessage)
+            let result = try predicate.satisfies(expr)
+            result.message.update(failureMessage: failureMessage)
+            return result.toBoolean(expectation: .toMatch)
         } catch let error {
             failureMessage.stringValue = "unexpected error thrown: <\(error)>"
             return false
@@ -66,10 +68,12 @@ public class NMBObjCBeCloseToMatcher: NSObject, NMBMatcher {
             return actualExpression() as? NMBDoubleConvertible
         })
         let expr = Expression(expression: actualBlock, location: location)
-        let matcher = beCloseTo(self._expected, within: self._delta)
+        let predicate = beCloseTo(self._expected, within: self._delta)
 
         do {
-            return try matcher.doesNotMatch(expr, failureMessage: failureMessage)
+            let result = try predicate.satisfies(expr)
+            result.message.update(failureMessage: failureMessage)
+            return result.toBoolean(expectation: .toNotMatch)
         } catch let error {
             failureMessage.stringValue = "unexpected error thrown: <\(error)>"
             return false
@@ -77,9 +81,9 @@ public class NMBObjCBeCloseToMatcher: NSObject, NMBMatcher {
     }
 
     @objc public var within: (CDouble) -> NMBObjCBeCloseToMatcher {
-        return ({ delta in
+        return { delta in
             return NMBObjCBeCloseToMatcher(expected: self._expected, within: delta)
-        })
+        }
     }
 }
 
