@@ -11,6 +11,11 @@ public func beGreaterThan<T: Comparable>(_ expectedValue: T?) -> Predicate<T> {
     }
 }
 
+public func ><T: Comparable>(lhs: Expectation<T>, rhs: T) {
+    lhs.to(beGreaterThan(rhs))
+}
+
+#if canImport(Darwin) || !compiler(>=5.1)
 /// A Nimble matcher that succeeds when the actual value is greater than the expected value.
 public func beGreaterThan(_ expectedValue: NMBComparable?) -> Predicate<NMBComparable> {
     let errorMessage = "be greater than <\(stringify(expectedValue))>"
@@ -22,13 +27,10 @@ public func beGreaterThan(_ expectedValue: NMBComparable?) -> Predicate<NMBCompa
     }
 }
 
-public func ><T: Comparable>(lhs: Expectation<T>, rhs: T) {
-    lhs.to(beGreaterThan(rhs))
-}
-
 public func > (lhs: Expectation<NMBComparable>, rhs: NMBComparable?) {
     lhs.to(beGreaterThan(rhs))
 }
+#endif
 
 #if canImport(Darwin)
 extension NMBObjCMatcher {
