@@ -13,6 +13,11 @@ public func beGreaterThanOrEqualTo<T: Comparable>(_ expectedValue: T?) -> Predic
     }
 }
 
+public func >=<T: Comparable>(lhs: Expectation<T>, rhs: T) {
+    lhs.to(beGreaterThanOrEqualTo(rhs))
+}
+
+#if canImport(Darwin) || !compiler(>=5.1)
 /// A Nimble matcher that succeeds when the actual value is greater than
 /// or equal to the expected value.
 public func beGreaterThanOrEqualTo<T: NMBComparable>(_ expectedValue: T?) -> Predicate<T> {
@@ -24,13 +29,10 @@ public func beGreaterThanOrEqualTo<T: NMBComparable>(_ expectedValue: T?) -> Pre
     }
 }
 
-public func >=<T: Comparable>(lhs: Expectation<T>, rhs: T) {
-    lhs.to(beGreaterThanOrEqualTo(rhs))
-}
-
 public func >=<T: NMBComparable>(lhs: Expectation<T>, rhs: T) {
     lhs.to(beGreaterThanOrEqualTo(rhs))
 }
+#endif
 
 #if canImport(Darwin)
 extension NMBObjCMatcher {
