@@ -18,13 +18,13 @@ extension Matcher {
     }
 
     var toClosure: (Expression<ValueType>, FailureMessage, Bool) throws -> Bool {
-        return ({ expr, msg, expectedResult in
+        return { expr, msg, expectedResult in
             if expectedResult {
                 return try self.matches(expr, failureMessage: msg)
             } else {
                 return try self.doesNotMatch(expr, failureMessage: msg)
             }
-        })
+        }
     }
 }
 
@@ -136,13 +136,14 @@ extension NSDate: TestOutputStringConvertible {
 @objc public protocol NMBComparable {
     func NMB_compare(_ otherObject: NMBComparable!) -> ComparisonResult
 }
-#else
+#elseif !compiler(>=5.1)
 // This should become obsolete once Corelibs Foundation adds Comparable conformance to NSNumber
 public protocol NMBComparable {
     func NMB_compare(_ otherObject: NMBComparable!) -> ComparisonResult
 }
 #endif
 
+#if canImport(Darwin) || !compiler(>=5.1)
 extension NSNumber: NMBComparable {
     public func NMB_compare(_ otherObject: NMBComparable!) -> ComparisonResult {
         // swiftlint:disable:next force_cast
@@ -155,3 +156,4 @@ extension NSString: NMBComparable {
         return compare(otherObject as! String)
     }
 }
+#endif
