@@ -3,11 +3,7 @@ import Foundation
 internal class NotificationCollector {
     private(set) var observedNotifications: [Notification]
     private let notificationCenter: NotificationCenter
-    #if canImport(Darwin)
-    private var token: AnyObject?
-    #else
     private var token: NSObjectProtocol?
-    #endif
 
     required init(notificationCenter: NotificationCenter) {
         self.notificationCenter = notificationCenter
@@ -23,19 +19,52 @@ internal class NotificationCollector {
     }
 
     deinit {
-        #if canImport(Darwin)
-            if let token = self.token {
-                self.notificationCenter.removeObserver(token)
-            }
-        #else
-            if let token = self.token as? AnyObject {
-                self.notificationCenter.removeObserver(token)
-            }
-        #endif
+        if let token = self.token {
+            self.notificationCenter.removeObserver(token)
+        }
     }
 }
 
 private let mainThread = pthread_self()
+
+public func postNotifications(
+    _ predicate: Predicate<[Notification]>,
+    fromNotificationCenter center: NotificationCenter = .default
+) -> Predicate<Any> {
+    _ = mainThread // Force lazy-loading of this value
+    let collector = NotificationCollector(notificationCenter: center)
+    collector.startObserving()
+    var once: Bool = false
+
+    return Predicate { actualExpression in
+        let collectorNotificationsExpression = Expression(
+            memoizedExpression: { _ in
+                return collector.observedNotifications
+            },
+            location: actualExpression.location,
+            withoutCaching: true
+        )
+
+        assert(pthread_equal(mainThread, pthread_self()) != 0, "Only expecting closure to be evaluated on main thread.")
+        if !once {
+            once = true
+            _ = try actualExpression.evaluate()
+        }
+
+        let actualValue: String
+        if collector.observedNotifications.isEmpty {
+            actualValue = "no notifications"
+        } else {
+            actualValue = "<\(stringify(collector.observedNotifications))>"
+        }
+
+        var result = try predicate.satisfies(collectorNotificationsExpression)
+        result.message = result.message.replacedExpectation { message in
+            return .expectedCustomValueTo(message.expectedMessage, actualValue)
+        }
+        return result
+    }
+}
 
 public func postNotifications<T>(
     _ notificationsMatcher: T,
